+++ HISTORY.rst
@@ -11,6 +11,7 @@ Upcoming
 - Flaky will no longer raise a UnicodeEncodeError for flaky tests which raise exceptions
   with non-ascii characters.
 - Flaky will no longer cause nose to report non-flaky test failures and errors twice.
+- Flaky now works with tests that are parametrized with py.test.
 
 
 2.1.1 (2015-05-22)